MultiLanguage-Invoice_Extractor-GeminiPro/requirements.txt
streamlit
langchain
google.generativeai
python-dotenv
PyPDF2
chromadb # vector embeddings/vector stores

